refactor: modularize room page and backend
