+++ frontend/pages/room/[roomId].tsx
@@ -4,6 +4,9 @@ import { useAuth } from '../../context/AuthContext';
 import SimplePeer from 'simple-peer';
 import { v4 as uuidv4 } from 'uuid';
 import { startActivityTracking, stopActivityTracking, updateActivityHeartbeat, getUserRoleInOffice } from '../../lib/firebase';
+import ChatPanel from "../../components/room/ChatPanel";
+import Whiteboard from "../../components/room/Whiteboard";
+import ControlPanel from "../../components/room/ControlPanel";
 
 interface PeerRef {
   id: string;
@@ -1605,532 +1608,44 @@ export default function Room() {
               )}
             </div>
           ) : (
-            /* Modern Whiteboard */
-            <div style={{ 
-              position: 'relative',
-              background: 'linear-gradient(135deg, #f8fafc 0%, #e2e8f0 100%)',
-              borderRadius: '20px',
-              padding: '24px',
-              boxShadow: '0 8px 32px rgba(0, 0, 0, 0.4)',
-              border: '2px solid rgba(255, 255, 255, 0.1)'
-            }}>
-              <canvas
-                ref={canvasRef}
-                width={800}
-                height={600}
-                style={{
-                  border: '2px solid #e2e8f0',
-                  borderRadius: '12px',
-                  cursor: 'crosshair',
-                  backgroundColor: 'white'
-                }}
-                onMouseDown={startDrawing}
-                onMouseMove={draw}
-                onMouseUp={stopDrawing}
-                onMouseLeave={stopDrawing}
-              />
-              <button
-                onClick={clearWhiteboard}
-                style={{
-                  position: 'absolute',
-                  top: '36px',
-                  right: '36px',
-                  padding: '10px 20px',
-                  background: 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)',
-                  color: 'white',
-                  border: 'none',
-                  borderRadius: '12px',
-                  cursor: 'pointer',
-                  fontSize: '14px',
-                  fontWeight: '600',
-                  boxShadow: '0 4px 12px rgba(239, 68, 68, 0.4)',
-                  transition: 'all 0.2s ease'
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.transform = 'translateY(-2px)';
-                  e.currentTarget.style.boxShadow = '0 6px 16px rgba(239, 68, 68, 0.5)';
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.transform = 'translateY(0)';
-                  e.currentTarget.style.boxShadow = '0 4px 12px rgba(239, 68, 68, 0.4)';
-                }}
-              >
-                Clear Board
-              </button>
-            </div>
+            <Whiteboard
+              show={showWhiteboard}
+              canvasRef={canvasRef}
+              startDrawing={startDrawing}
+              draw={draw}
+              stopDrawing={stopDrawing}
+              clearWhiteboard={clearWhiteboard}
+            />
           )}
         </div>
 
         {/* Chat Panel */}
-        {showChat && (
-        <div style={{
-            position: isMobile ? 'fixed' : 'relative',
-            top: isMobile ? '0' : 'auto',
-            left: isMobile ? '0' : 'auto',
-            right: isMobile ? '0' : 'auto',
-            bottom: isMobile ? '0' : 'auto',
-            width: isMobile ? '100%' : '300px',
-            height: isMobile ? '100vh' : 'auto',
-          backgroundColor: '#202124',
-            borderLeft: isMobile ? 'none' : '1px solid #5f6368',
-          display: 'flex',
-            flexDirection: 'column',
-            zIndex: isMobile ? 1000 : 'auto'
-        }}>
-          <div style={{
-              padding: isMobile ? '16px 16px 12px 16px' : '16px',
-            borderBottom: '1px solid #5f6368',
-            fontWeight: '600',
-              fontSize: '16px',
-              display: 'flex',
-              justifyContent: 'space-between',
-              alignItems: 'center'
-            }}>
-              <span>Chat</span>
-              <button
-                onClick={() => setShowChat(false)}
-                style={{
-                  background: 'none',
-                  border: 'none',
-                  color: 'white',
-                  fontSize: '20px',
-                  cursor: 'pointer',
-                  padding: '4px'
-                }}
-              >
-                ✕
-              </button>
-          </div>
-          
-          <div style={{
-            flex: 1,
-            padding: '16px',
-            overflowY: 'auto',
-            display: 'flex',
-            flexDirection: 'column',
-            gap: '8px'
-          }}>
-            {messages.map((msg, idx) => (
-              <div key={idx} style={{
-                padding: '8px 12px',
-                  backgroundColor: msg.sender === myId.current ? '#0F9D58' : '#3c4043',
-                borderRadius: '12px',
-                fontSize: '14px',
-                wordBreak: 'break-word'
-              }}>
-                <div style={{ fontWeight: '600', marginBottom: '2px' }}>
-                  {msg.sender === myId.current ? 'You' : (participants.find(p => p.id === msg.sender)?.name || 'Unknown')}
-                </div>
-                <div>{msg.text}</div>
-                {msg.timestamp && (
-                  <div style={{ fontSize: '11px', opacity: 0.7, marginTop: '2px' }}>
-                    {msg.timestamp}
-                  </div>
-                )}
-              </div>
-            ))}
-          </div>
-          
-          <div style={{
-            padding: '16px',
-            borderTop: '1px solid #5f6368',
-            display: 'flex',
-              gap: '8px',
-              paddingBottom: isMobile ? '24px' : '16px'
-          }}>
-            <input
-              type="text"
-              value={newMsg}
-              onChange={(e) => setNewMsg(e.target.value)}
-              onKeyPress={(e) => e.key === 'Enter' && sendMessage()}
-              placeholder="Type a message..."
-              style={{
-                flex: 1,
-                  padding: '12px',
-                backgroundColor: '#3c4043',
-                border: '1px solid #5f6368',
-                borderRadius: '8px',
-                color: 'white',
-                fontSize: '14px',
-                outline: 'none'
-              }}
-            />
-            <button
-              onClick={sendMessage}
-              disabled={!newMsg.trim()}
-              style={{
-                  padding: '12px 16px',
-                  backgroundColor: newMsg.trim() ? '#0F9D58' : '#5f6368',
-                color: 'white',
-                border: 'none',
-                borderRadius: '8px',
-                cursor: newMsg.trim() ? 'pointer' : 'not-allowed',
-                  fontSize: '14px',
-                  fontWeight: '500'
-              }}
-            >
-              Send
-            </button>
-          </div>
-        </div>
-        )}
-      </div>
-
-      {/* Floating Control Buttons */}
-      <div style={{
-        position: 'fixed',
-        bottom: isMobile ? '24px' : '32px',
-        left: '50%',
-        transform: 'translateX(-50%)',
-        zIndex: 100,
-        pointerEvents: 'none'
-      }}>
-        {/* Control Container */}
-        <div style={{ 
-          display: 'flex', 
-          justifyContent: 'center',
-          alignItems: 'center',
-          gap: isMobile ? '16px' : '20px',
-          background: 'rgba(0, 0, 0, 0.8)',
-          backdropFilter: 'blur(20px)',
-          borderRadius: '50px',
-          padding: isMobile ? '12px 20px' : '16px 24px',
-          border: '1px solid rgba(255, 255, 255, 0.1)',
-          boxShadow: '0 8px 32px rgba(0, 0, 0, 0.4)',
-          pointerEvents: 'auto'
-        }}>
-          {/* Left Controls Group */}
-          <div style={{ 
-            display: 'flex', 
-            gap: isMobile ? '12px' : '16px',
-            alignItems: 'center'
-          }}>
-            {/* Microphone */}
-            <button 
-              onClick={() => {
-                if (streamRef.current) {
-                  const audioTrack = streamRef.current.getAudioTracks()[0];
-                  if (audioTrack) {
-                    audioTrack.enabled = !audioTrack.enabled;
-                    setPeers([...peersRef.current || []]);
-                  }
-                }
-              }}
-              disabled={!streamRef.current}
-              style={{
-                width: isMobile ? '56px' : '52px',
-                height: isMobile ? '56px' : '52px',
-                borderRadius: '50%',
-                background: streamRef.current?.getAudioTracks()[0]?.enabled !== false 
-                  ? 'linear-gradient(135deg, rgba(255, 255, 255, 0.15) 0%, rgba(255, 255, 255, 0.05) 100%)'
-                  : 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)',
-                border: streamRef.current?.getAudioTracks()[0]?.enabled !== false 
-                  ? '2px solid rgba(255, 255, 255, 0.2)'
-                  : '2px solid #ef4444',
-                color: 'white',
-                cursor: 'pointer',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-                boxShadow: streamRef.current?.getAudioTracks()[0]?.enabled !== false 
-                  ? '0 8px 24px rgba(0, 0, 0, 0.3)'
-                  : '0 8px 24px rgba(239, 68, 68, 0.4)',
-                outline: 'none',
-                transform: 'scale(1)',
-                backdropFilter: 'blur(10px)'
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.transform = 'scale(1.1)';
-                e.currentTarget.style.boxShadow = streamRef.current?.getAudioTracks()[0]?.enabled !== false 
-                  ? '0 12px 32px rgba(0, 0, 0, 0.4)'
-                  : '0 12px 32px rgba(239, 68, 68, 0.6)';
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.transform = 'scale(1)';
-                e.currentTarget.style.boxShadow = streamRef.current?.getAudioTracks()[0]?.enabled !== false 
-                  ? '0 8px 24px rgba(0, 0, 0, 0.3)'
-                  : '0 8px 24px rgba(239, 68, 68, 0.4)';
-              }}
-              title={streamRef.current?.getAudioTracks()[0]?.enabled !== false ? 'Mute microphone (M)' : 'Unmute microphone (M)'}
-            >
-              <svg width={isMobile ? "24" : "20"} height={isMobile ? "24" : "20"} viewBox="0 0 24 24" fill="currentColor">
-                {streamRef.current?.getAudioTracks()[0]?.enabled !== false ? (
-                  <path d="M12 14c1.66 0 3-1.34 3-3V5c0-1.66-1.34-3-3-3S9 3.34 9 5v6c0 1.66 1.34 3 3 3zm5.3-3c0 3-2.54 5.1-5.3 5.1S6.7 14 6.7 11H5c0 3.41 2.72 6.23 6 6.72V21h2v-3.28c3.28-.49 6-3.31 6-6.72h-1.7z"/>
-                ) : (
-                  <path d="M19 11h-1.7c0 .74-.16 1.43-.43 2.05l1.23 1.23c.56-.98.9-2.09.9-3.28zm-4.02.17c0-.06.02-.11.02-.17V5c0-1.66-1.34-3-3-3S9 3.34 9 5v.18l5.98 5.99zM4.27 3L3 4.27l6.01 6.01V11c0 1.66 1.33 3 2.99 3 .22 0 .44-.03.65-.08l1.66 1.66c-.71.33-1.5.52-2.31.52-2.76 0-5.3-2.1-5.3-5.1H5c0 3.41 2.72 6.23 6 6.72V21h2v-3.28c.91-.13 1.77-.45 2.54-.9L19.73 21 21 19.73 4.27 3z"/>
-                )}
-              </svg>
-            </button>
-
-            {/* Camera */}
-            <button 
-              onClick={() => {
-                if (streamRef.current) {
-                  const videoTrack = streamRef.current.getVideoTracks()[0];
-                  if (videoTrack) {
-                    videoTrack.enabled = !videoTrack.enabled;
-                    setPeers([...peersRef.current || []]);
-                  }
-                }
-              }}
-              disabled={!streamRef.current}
-              style={{
-                width: isMobile ? '56px' : '52px',
-                height: isMobile ? '56px' : '52px',
-                borderRadius: '50%',
-                background: streamRef.current?.getVideoTracks()[0]?.enabled !== false 
-                  ? 'linear-gradient(135deg, rgba(255, 255, 255, 0.15) 0%, rgba(255, 255, 255, 0.05) 100%)'
-                  : 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)',
-                border: streamRef.current?.getVideoTracks()[0]?.enabled !== false 
-                  ? '2px solid rgba(255, 255, 255, 0.2)'
-                  : '2px solid #ef4444',
-                color: 'white',
-                cursor: 'pointer',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-                boxShadow: streamRef.current?.getVideoTracks()[0]?.enabled !== false 
-                  ? '0 8px 24px rgba(0, 0, 0, 0.3)'
-                  : '0 8px 24px rgba(239, 68, 68, 0.4)',
-                outline: 'none',
-                transform: 'scale(1)',
-                backdropFilter: 'blur(10px)'
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.transform = 'scale(1.1)';
-                e.currentTarget.style.boxShadow = streamRef.current?.getVideoTracks()[0]?.enabled !== false 
-                  ? '0 12px 32px rgba(0, 0, 0, 0.4)'
-                  : '0 12px 32px rgba(239, 68, 68, 0.6)';
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.transform = 'scale(1)';
-                e.currentTarget.style.boxShadow = streamRef.current?.getVideoTracks()[0]?.enabled !== false 
-                  ? '0 8px 24px rgba(0, 0, 0, 0.3)'
-                  : '0 8px 24px rgba(239, 68, 68, 0.4)';
-              }}
-              title={streamRef.current?.getVideoTracks()[0]?.enabled !== false ? 'Turn off camera (C)' : 'Turn on camera (C)'}
-            >
-              <svg width={isMobile ? "24" : "20"} height={isMobile ? "24" : "20"} viewBox="0 0 24 24" fill="currentColor">
-                {streamRef.current?.getVideoTracks()[0]?.enabled !== false ? (
-                  <path d="M17 10.5V7c0-.55-.45-1-1-1H4c-.55 0-1 .45-1 1v10c0 .55.45 1 1 1h12c.55 0 1-.45 1-1v-3.5l4 4v-11l-4 4z"/>
-                ) : (
-                  <path d="M21 6.5l-4 4V7c0-.55-.45-1-1-1H9.82L21 17.18V6.5zM3.27 2L2 3.27 4.73 6H4c-.55 0-1 .45-1 1v10c0 .55.45 1 1 1h12c.21 0 .39-.08.55-.18L19.73 21 21 19.73 3.27 2z"/>
-                )}
-              </svg>
-            </button>
-          </div>
-
-          {/* Center - End Call Button */}
-          <button 
-            onClick={() => router.push(`/office/${officeId}`)}
-            style={{
-              width: isMobile ? '64px' : '60px',
-              height: isMobile ? '64px' : '60px',
-              borderRadius: '50%',
-              background: 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)',
-              border: 'none',
-              color: 'white',
-              cursor: 'pointer',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-              transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-              boxShadow: '0 8px 32px rgba(239, 68, 68, 0.5)',
-              outline: 'none',
-              transform: 'scale(1)',
-              position: 'relative'
-            }}
-            onMouseEnter={(e) => {
-              e.currentTarget.style.transform = 'scale(1.1)';
-              e.currentTarget.style.boxShadow = '0 12px 40px rgba(239, 68, 68, 0.7)';
-            }}
-            onMouseLeave={(e) => {
-              e.currentTarget.style.transform = 'scale(1)';
-              e.currentTarget.style.boxShadow = '0 8px 32px rgba(239, 68, 68, 0.5)';
-            }}
-            title="Leave call (ESC)"
-          >
-            <svg width={isMobile ? "28" : "24"} height={isMobile ? "28" : "24"} viewBox="0 0 24 24" fill="currentColor">
-              <path d="M12 9c-1.6 0-3.15.25-4.6.72v3.1c0 .39-.23.74-.56.9-.98.49-1.87 1.12-2.66 1.85-.18.18-.43.28-.7.28-.28 0-.53-.11-.71-.29L.29 13.08c-.18-.17-.29-.42-.29-.7 0-.28.11-.53.29-.71C3.34 8.78 7.46 7 12 7s8.66 1.78 11.71 4.67c.18.18.29.43.29.71 0 .28-.11.53-.29.7l-2.48 2.48c-.18.18-.43.29-.71.29-.27 0-.52-.1-.7-.28-.79-.73-1.68-1.36-2.66-1.85-.33-.16-.56-.51-.56-.9v-3.1C15.15 9.25 13.6 9 12 9z"/>
-            </svg>
-          </button>
-
-          {/* Right Controls Group */}
-          <div style={{ 
-            display: 'flex', 
-            gap: isMobile ? '12px' : '16px',
-            alignItems: 'center'
-          }}>
-            {/* Screen Share */}
-            {!isMobile && (
-              <button 
-                onClick={toggleScreenShare}
-                disabled={!streamRef.current}
-                style={{
-                  width: '52px',
-                  height: '52px',
-                  borderRadius: '50%',
-                  background: isScreenSharing 
-                    ? 'linear-gradient(135deg, #3b82f6 0%, #1d4ed8 100%)'
-                    : 'linear-gradient(135deg, rgba(255, 255, 255, 0.15) 0%, rgba(255, 255, 255, 0.05) 100%)',
-                  border: isScreenSharing 
-                    ? '2px solid #3b82f6'
-                    : '2px solid rgba(255, 255, 255, 0.2)',
-                  color: 'white',
-                  cursor: 'pointer',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center',
-                  transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-                  boxShadow: isScreenSharing 
-                    ? '0 8px 24px rgba(59, 130, 246, 0.4)'
-                    : '0 8px 24px rgba(0, 0, 0, 0.3)',
-                  outline: 'none',
-                  transform: 'scale(1)',
-                  backdropFilter: 'blur(10px)'
-                }}
-                onMouseEnter={(e) => {
-                  e.currentTarget.style.transform = 'scale(1.1)';
-                  e.currentTarget.style.boxShadow = isScreenSharing 
-                    ? '0 12px 32px rgba(59, 130, 246, 0.6)'
-                    : '0 12px 32px rgba(0, 0, 0, 0.4)';
-                }}
-                onMouseLeave={(e) => {
-                  e.currentTarget.style.transform = 'scale(1)';
-                  e.currentTarget.style.boxShadow = isScreenSharing 
-                    ? '0 8px 24px rgba(59, 130, 246, 0.4)'
-                    : '0 8px 24px rgba(0, 0, 0, 0.3)';
-                }}
-                title={isScreenSharing ? 'Stop sharing' : 'Share screen'}
-              >
-                <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor">
-                  <path d="M20 18c1.1 0 1.99-.9 1.99-2L22 6c0-1.1-.9-2-2-2H4c-1.1 0-2 .9-2 2v10c0 1.1.9 2 2 2H0v2h24v-2h-4zM4 6h16v10H4V6z"/>
-                </svg>
-              </button>
-            )}
-
-            {/* Chat */}
-            <button 
-              onClick={() => setShowChat(!showChat)}
-              style={{
-                width: isMobile ? '56px' : '52px',
-                height: isMobile ? '56px' : '52px',
-                borderRadius: '50%',
-                background: showChat 
-                  ? 'linear-gradient(135deg, #3b82f6 0%, #1d4ed8 100%)'
-                  : 'linear-gradient(135deg, rgba(255, 255, 255, 0.15) 0%, rgba(255, 255, 255, 0.05) 100%)',
-                border: showChat 
-                  ? '2px solid #3b82f6'
-                  : '2px solid rgba(255, 255, 255, 0.2)',
-                color: 'white',
-                cursor: 'pointer',
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'center',
-                transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-                boxShadow: showChat 
-                  ? '0 8px 24px rgba(59, 130, 246, 0.4)'
-                  : '0 8px 24px rgba(0, 0, 0, 0.3)',
-                outline: 'none',
-                transform: 'scale(1)',
-                position: 'relative',
-                backdropFilter: 'blur(10px)'
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.transform = 'scale(1.1)';
-                e.currentTarget.style.boxShadow = showChat 
-                  ? '0 12px 32px rgba(59, 130, 246, 0.6)'
-                  : '0 12px 32px rgba(0, 0, 0, 0.4)';
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.transform = 'scale(1)';
-                e.currentTarget.style.boxShadow = showChat 
-                  ? '0 8px 24px rgba(59, 130, 246, 0.4)'
-                  : '0 8px 24px rgba(0, 0, 0, 0.3)';
-              }}
-              title="Toggle chat"
-            >
-              <svg width={isMobile ? "24" : "20"} height={isMobile ? "24" : "20"} viewBox="0 0 24 24" fill="currentColor">
-                <path d="M20 2H4c-1.1 0-1.99.9-1.99 2L2 22l4-4h14c1.1 0 2-.9 2-2V4c0-1.1-.9-2-2-2zm-2 12H6v-2h12v2zm0-3H6V9h12v2zm0-3H6V6h12v2z"/>
-              </svg>
-              {messages.length > 0 && !showChat && (
-                <div style={{
-                  position: 'absolute',
-                  top: '-4px',
-                  right: '-4px',
-                  width: '20px',
-                  height: '20px',
-                  borderRadius: '50%',
-                  backgroundColor: '#EA4335',
-                  fontSize: '11px',
-                  fontWeight: '600',
-                  display: 'flex',
-                  alignItems: 'center',
-                  justifyContent: 'center'
-                }}>
-                  {messages.length > 9 ? '9+' : messages.length}
-                </div>
-              )}
-            </button>
-
-            {/* More Options */}
-          <button 
-            onClick={() => setShowWhiteboard(!showWhiteboard)}
-            style={{
-                width: isMobile ? '52px' : '48px',
-                height: isMobile ? '52px' : '48px',
-              borderRadius: '50%',
-                backgroundColor: showWhiteboard ? '#8E24AA' : 'rgba(255, 255, 255, 0.1)',
-                border: `2px solid ${showWhiteboard ? '#8E24AA' : 'rgba(255, 255, 255, 0.2)'}`,
-              color: 'white',
-              cursor: 'pointer',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-                transition: 'all 0.3s cubic-bezier(0.4, 0, 0.2, 1)',
-                boxShadow: '0 4px 12px rgba(0, 0, 0, 0.15)',
-                outline: 'none',
-                transform: 'scale(1)'
-              }}
-              onMouseEnter={(e) => {
-                e.currentTarget.style.transform = 'scale(1.05)';
-                e.currentTarget.style.boxShadow = '0 6px 16px rgba(0, 0, 0, 0.2)';
-              }}
-              onMouseLeave={(e) => {
-                e.currentTarget.style.transform = 'scale(1)';
-                e.currentTarget.style.boxShadow = '0 4px 12px rgba(0, 0, 0, 0.15)';
-              }}
-              title={showWhiteboard ? 'Hide whiteboard' : 'Open whiteboard'}
-            >
-              <svg width={isMobile ? "24" : "18"} height={isMobile ? "24" : "18"} viewBox="0 0 24 24" fill="currentColor">
-                <path d="M22 5v2h-3v3h-2V7h-3V5h3V2h2v3h3zM6 7h9v2H6V7zm0 4h9v2H6v-2zm0 4h6v2H6v-2zm10-1.5V14h3v-2h-3v-.5zM4 2H2v18l4-4h14c1.1 0 2-.9 2-2V5h-2v9H6.17L4 16.17V2z"/>
-              </svg>
-          </button>
-        </div>
-      </div>
-
-        {/* Connection Status */}
-        <div style={{
-          position: 'absolute',
-          top: '8px',
-          left: '50%',
-          transform: 'translateX(-50%)',
-          fontSize: '12px',
-          color: 'rgba(255, 255, 255, 0.7)',
-          display: 'flex',
-          alignItems: 'center',
-          gap: '6px'
-        }}>
-          <div style={{
-            width: '8px',
-            height: '8px',
-            borderRadius: '50%',
-            backgroundColor: isConnected ? '#0F9D58' : '#EA4335'
-          }} />
-          {connectionStatus}
-        </div>
+        <ChatPanel
+          isMobile={isMobile}
+          showChat={showChat}
+          setShowChat={setShowChat}
+          messages={messages}
+          newMsg={newMsg}
+          setNewMsg={setNewMsg}
+          sendMessage={sendMessage}
+          participants={participants}
+          myId={myId}
+        />
+        <ControlPanel
+          isMobile={isMobile}
+          streamRef={streamRef}
+          toggleScreenShare={toggleScreenShare}
+          isScreenSharing={isScreenSharing}
+          showChat={showChat}
+          setShowChat={setShowChat}
+          showWhiteboard={showWhiteboard}
+          setShowWhiteboard={setShowWhiteboard}
+          router={router}
+          officeId={officeId}
+          messages={messages}
+          isConnected={isConnected}
+          connectionStatus={connectionStatus}
+        />
       </div>
 
       {/* Profile Modal */}
